+++ src/app/static/src/tests/root/getters.test.ts
@@ -9,7 +9,7 @@ import {
     mockLoadState,
     mockMetadataState,
     mockModelCalibrateState,
-    mockModelOptionsState, mockModelOutputState,
+    mockModelOptionsState,
     mockModelRunState, mockProjectOutputState, mockProjectsState,
     mockRootState, mockSurveyAndProgramState
 } from "../mocks";
@@ -399,28 +399,83 @@ describe("root getters", () => {
 
     it(`can get serialized project states`, () => {
         const projectStates = () => {
-            return mockRootState({
-                baseline: mockBaselineState({
-                        pjnz: {filename: "pjnz", hash: "pjnzHash"} as any,
-                        population: {filename: "population", hash: "populationHash"} as any,
-                        shape: {filename: "shape", hash: "shapeHash"} as any
-                    }),
-                surveyAndProgram: mockSurveyAndProgramState({
-                    warnings: surveyAndProgramWarnings,
-                    anc: {filename: "anc", hash: "ancHash"} as any,
-                    program: {filename: "program", hash: "programHash"} as any,
-                    survey: {filename: "survey", hash: "surveyHash"} as any
-                }),
-                modelOptions: mockModelOptionsState({options: {"test": "options"}}),
-                modelRun: mockModelRunState({modelRunId: "modelRunId"}),
-                modelCalibrate: mockModelCalibrateState({calibrateId: "calibrateId", options: {"test": "options"}}),
-                hintrVersion: mockHintrVersionState({hintrVersion: {hintr: "1.0.0", naomi: "2.0.0", rrq: "1.1.1"}})
-            })
+            return mockRootState(projectStateTestData({
+                projects: mockProjectsState({
+                    currentProject:
+                        {
+                            name: "My project 123",
+                            note: "These are my project notes",
+                            versions: [
+                                {
+                                    note: "Notes specific to this version",
+                                    updated: "2022-06-19T13:56:19.280Z",
+                                    versionNumber: "1"
+                                },
+                                {
+                                    note: "Notes from the first version",
+                                    updated: "2022-06-19T13:56:19.280Z",
+                                    versionNumber: "2"
+                                }
+                            ]
+                        } as any
+                })
+            }))
         }
+
         const rootState = projectStates()
 
         const result = getters.projectState(rootState, null, projectStates() as any, null)
-        expect(result).toEqual(mockProjectOutputState())
-    })
 
+        const millis = Date.UTC(2022, 6, 19, 13, 56, 19)
+
+        const date = new Date(millis)
+
+        const updated = `2022/06/${date.getDate()} ${date.getHours()}:${date.getMinutes()}:${date.getSeconds()}`
+
+        expect(result).toStrictEqual(mockProjectOutputState({
+            notes: {
+                project_notes: {
+                    name: "My project 123",
+                    updated: updated,
+                    note: "These are my project notes"
+                },
+                version_notes: [
+                    {
+                        name: "My project 123-v1",
+                        updated: updated,
+                        note: "Notes specific to this version"
+                    },
+                    {
+                        name: "My project 123-v2",
+                        updated: updated,
+                        note: "Notes from the first version"
+                    }
+                ]
+            } as any
+        }))
+    })
 })
+
+const projectStateTestData = (props: Partial<any> = {}) => {
+    const surveyAndProgramWarnings: Warning[] = [
+        {text: "survey and program test", locations: ["review_inputs"]}
+    ]
+    return {
+        baseline: mockBaselineState({
+            pjnz: {filename: "pjnz", hash: "pjnzHash"} as any,
+            population: {filename: "population", hash: "populationHash"} as any,
+            shape: {filename: "shape", hash: "shapeHash"} as any
+        }),
+        surveyAndProgram: mockSurveyAndProgramState({
+            warnings: surveyAndProgramWarnings,
+            anc: {filename: "anc", hash: "ancHash"} as any,
+            program: {filename: "program", hash: "programHash"} as any,
+            survey: {filename: "survey", hash: "surveyHash"} as any
+        }),
+        modelOptions: mockModelOptionsState({options: {"test": "options"}}),
+        modelRun: mockModelRunState({modelRunId: "modelRunId"}),
+        modelCalibrate: mockModelCalibrateState({calibrateId: "calibrateId", options: {"test": "options"}}),
+        hintrVersion: mockHintrVersionState({hintrVersion: {hintr: "1.0.0", naomi: "2.0.0", rrq: "1.1.1"}}),
+        ...props
+    }
+}