+++ NEWS.md
@@ -1,3 +1,15 @@
+# hint 2.2.1
+
+* Bug: Fix bug backwards compatibility calibrate options
+
+# hint 2.2.0
+
+* Include project notes in output zip
+
+# hint 2.1.1
+
+* Bug: hint tries to pull resources it does not have access to
+
 # hint 2.1.0
 
 * Use commas as thousands separators for plain numbers in the input time series table